+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 1.7.2 / ####-##-##
 ==================
 
+  * Changed browser.assert.url error message when internal error happens
+
 1.7.1 / 2019-01-08
 ==================