+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
   * WaitUntilEnabled
   * Changed browser.assert.url error message when internal error happens
+  * Fixed unexpected behaviour on closing a browser that has never been opened
 
 1.7.1 / 2019-01-08
 ==================